+++ incubator-access/src/main/java/com/michealyang/model/houseSpy/domain/LJHouse.java
@@ -10,7 +10,7 @@ public class LJHouse {
     private float area; //面积
     private String houseType;   //房型
     private String community;   //小区
-    private int finalTotal; //最终价格，房源下架以后展示
+    private float finalTotal; //最终价格，房源下架以后展示
     private int offShelf;   //是否已经下架
     private String imgs;    //房源图片
     private String url; //房源链接
@@ -83,11 +83,11 @@ public class LJHouse {
         this.offShelf = offShelf;
     }
 
-    public int getFinalTotal() {
+    public float getFinalTotal() {
         return finalTotal;
     }
 
-    public void setFinalTotal(int finalTotal) {
+    public void setFinalTotal(float finalTotal) {
         this.finalTotal = finalTotal;
     }
